[REEF-1687] Throw exception if task dispose throws exception after Call()
Currently in TaskRuntime, when task Dispose in Close method throws exception, we throw exception.
But when task Dispose in main flow (after Call method) throws exception, we log it but don't re-throw.
This behavior is inconsistent. It also results in random failures of TestFailMapperTasksOnDispose,
because the test depends on which code path is quicker and where the task Dispose is called.

When task Dispose throws exception, that means some resources are not released properly.
For fault tolerance work, even if the task returns from the Call(), either having really completed
the task or having been closed by driver, if we want to reuse the evaluator to resubmit a task on it,
we can't afford that. So the correct way is to throw exception to make evaluator fail,
same as what we do in Close method.

This change:
 * updates TaskRuntime to throw exception in both code paths.
 * updates TestFailMapperTasksOnDispose to account for the time-sensitive nature of the test

JIRA:
  [REEF-1687](https://issues.apache.org/jira/browse/REEF-1687)

Pull request:
  This closes #1203