[REEF-1511] Timeout for Task shutdown during IMRU recovery
During IMRU FT recovery, sometime, the tasks that supposed to close by driver don't send event back, causing the system hang. This PR adds the timeout for tasks closed by driver, and kill the evaluator if timeout.

The average closing time is recorded in the TaskManager. This number is a reference for define the time out on the fly. In case the average number is accessed before the average data is accumulated, or the average number is too low in some scenarios, we also have a configurable MinTaskWaitingForCloseTimeout to ensure the driver wait for enough time before killing the evaluator.

JIRA: [REEF-1511](https://issues.apache.org/jira/browse/REEF-1511)
This closes  #1201