[REEF-1778] Ensure ITask.Dispose() is executed
This moves the reporting of the result of a task *after* the call to `Dispose`
on that task to ensure execution of that method.

To do so, I made the calls to the task start and stop handlers explicit, and
kept track of the exceptions thrown and need to send a result in the
`TaskRuntime`.

Also, this fixes a bug in `TaskRuntimeTests` which did not wait for the
spawned thread to finish before validation of the test results.

JIRA:
  [REEF-1778](https://issues.apache.org/jira/browse/REEF-1778)

Pull Request:
  This closes #1293