[REEF-1578] Introduce DotNet build files for Utilities and Tang
This check-in starts the conversion of existing VS projects to the new dotnet
project format supported by Visual Studio 2017 and the dotnet command line.

This starts with conversion of the Utilities and Tang projects, providing a
`.DotNet.` project file that sits side-by-dside with the existing .net 4.5
project. This allows the existing projects to still be used during the
transition to the new project format.

With the new project format, a single project supports building multiple dotnet
frameworks. This provides the framework that will be followed with the other
projects. We are targeting `.net451`, `.net46` and `.netstarndard`/`.netcore`
where possible as there are some projects that cannot convert to .net core yet
(Tang is one example of this).

Since Utilities now targets .netstandard 2.0 - this addresses REEF-1578.

To build the new projects, the dotnet core cli tools will be required.
Details on this provided in the BUILD.md file. There is also a side-by-side
DotNet solution file that is used to build all migrated DotNet projects.

JIRA:
  [REEF-1578](https://issues.apache.org/jira/browse/REEF-1578)

Pull request:
  This closes #1279