[REEF-1950] Better test reporting for REEF.NET
This adds an API to write functional tests of REEF. The main API is
`ITestRunner` through which a Test can be submitted. In the Driver of a test,
one can use an injected instance of `Assert` to record the passing or faling of
tests.

A `ITestRunner` is obtained via the `TestRunnerFactory` which uses the same
environment variables used by the Java tests to determine which TestRunner to
instantiate.

A good way to understand the new functionality is the new functional test,
`TestTestFramework`.

Known issues:

  * The `IAssert` API is spartan.
  * There is only an implementation for the local `TestRunner`.

JIRA:
  [REEF-1950](https://issues.apache.org/jira/browse/REEF-1950)

Pull Request:
  This closes #1412