[REEF-1948] Add IREEFClient.NewJobRequestBuilder
This adds the new API to `IREEFClient`.

To ease implementation and foster re-use, this also adds the new class
`JobRequestBuilderFactory`. It is used on all implementations of
`IREEFClient` to implement the new method.

Lastly, this changes `HelloREEF` to use the new API to demonstrate it to
work.

JIRA:
  [REEF-1948](https://issues.apache.org/jira/browse/REEF-1948)

Pull Request:
  This closes #